+++ asyncqlio/backends/base.py
@@ -58,6 +58,13 @@ class BaseDialect:
         return False
 
     @property
+    def has_truncate(self) -> bool:
+        """
+        Returns TRUE if this dialect has TRUNCATE.
+        """
+        return False
+
+    @property
     def lastval_method(self):
         """
         The last value method for a dialect. For example, in PostgreSQL this is LASTVAL();