+++ asyncqlio/backends/postgresql/__init__.py
@@ -48,3 +48,7 @@ class PostgresqlDialect(BaseDialect):
     @property
     def has_default(self):
         return True
+
+    @property
+    def has_truncate(self):
+        return True